Rename LoyalVisitors to ReturningVisitors.
