+++ assets/js/modules/analytics-4/components/widgets/index.js
@@ -20,11 +20,11 @@ export { default as ConnectGA4CTATileWidget } from './ConnectGA4CTATileWidget';
 export { default as ConnectGA4CTAWidget } from './ConnectGA4CTAWidget';
 export { default as EngagedTrafficSourceWidget } from './EngagedTrafficSourceWidget';
 export { default as LeastEngagingPagesWidget } from './LeastEngagingPagesWidget';
-export { default as LoyalVisitorsWidget } from './LoyalVisitorsWidget';
 export { default as MostEngagingPagesWidget } from './MostEngagingPagesWidget';
 export { default as NewVisitorsWidget } from './NewVisitorsWidget';
 export { default as PopularContentWidget } from './PopularContentWidget';
 export { default as PopularProductsWidget } from './PopularProductsWidget';
+export { default as ReturningVisitorsWidget } from './ReturningVisitorsWidget';
 export { default as TopCitiesWidget } from './TopCitiesWidget';
 export { default as TopCountriesWidget } from './TopCountriesWidget';
 export { default as TopTrafficSourceWidget } from './TopTrafficSourceWidget';
